fix: review user card

+++ lib/screens/products/product_review.dart
@@ -26,45 +26,42 @@ class ProductReviewScreen extends GetView<ProductDetailController> {
         }
         final reviewsModel = controller.reviews.value;
         if (reviewsModel!.reviews.isEmpty) {
-          return Center(
-              child: Text("No data",
-                  style: Theme.of(context).textTheme.headlineMedium));
+          return Center(child: Text("No data", style: Theme.of(context).textTheme.headlineMedium));
         }
 
         return SingleChildScrollView(
-                child: Padding(
-                padding: const EdgeInsets.all(CustomSizes.defaultSpace),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(t.screens.productReview.subtitle),
-                    const SizedBox(height: CustomSizes.spaceBtwItems),
-                    // Rating Summary
-                    RatingSummary(
-                      counter: reviewsModel.reviews.length,
-                      average: reviewsModel.averageRating,
-                      showAverage: true,
-                      counterFiveStars: reviewsModel.count["5"]!,
-                      counterFourStars: reviewsModel.count["4"]!,
-                      counterThreeStars: reviewsModel.count["3"]!,
-                      counterTwoStars: reviewsModel.count["2"]!,
-                      counterOneStars: reviewsModel.count["1"]!,
-                      label: t.screens.productDetails.rating,
-                    ),
-                    const SizedBox(height: CustomSizes.spaceBtwSections),
-                    // User Review List
-                    ListView.builder(
-                        shrinkWrap: true,
-                        physics: const NeverScrollableScrollPhysics(),
-                        itemCount: reviewsModel.reviews.length,
-                        itemBuilder: (_, index) => UserReviewCard(
-                              review: reviewsModel.reviews[index],
-                              averageRating:
-                                  reviewsModel.averageRating,
-                            ))
-                  ],
-                ),
-              ));
+            child: Padding(
+          padding: const EdgeInsets.all(CustomSizes.defaultSpace),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(t.screens.productReview.subtitle),
+              const SizedBox(height: CustomSizes.spaceBtwItems),
+              // Rating Summary
+              RatingSummary(
+                counter: reviewsModel.reviews.length,
+                average: reviewsModel.averageRating.toDouble(),
+                showAverage: true,
+                counterFiveStars: reviewsModel.count["5"]!,
+                counterFourStars: reviewsModel.count["4"]!,
+                counterThreeStars: reviewsModel.count["3"]!,
+                counterTwoStars: reviewsModel.count["2"]!,
+                counterOneStars: reviewsModel.count["1"]!,
+                label: t.screens.productDetails.rating,
+              ),
+              const SizedBox(height: CustomSizes.spaceBtwSections),
+              // User Review List
+              ListView.builder(
+                  shrinkWrap: true,
+                  physics: const NeverScrollableScrollPhysics(),
+                  itemCount: reviewsModel.reviews.length,
+                  itemBuilder: (_, index) => UserReviewCard(
+                        review: reviewsModel.reviews[index],
+                        averageRating: reviewsModel.averageRating.toDouble(),
+                      ))
+            ],
+          ),
+        ));
       }),
     );
   }
